Tried to move project to old folder because of a building bug. Changed previous Pokémon script bug. Added possible loading icons.
Tried to move project to a new folder to fix the flutter building issue but realized it was a problem with the flutter sdk and the dart sdk as they were apparently both corrupted due to iCloud storage issues.

Changed the previous evolution script as some pokemon would have previous evolutions that were introduced in future generations and would instead show the Pokemon immediately before it in the Kanto dex.

Added possible loading Pokéball icons courtesy of Henry Lim on Dribbble and Travis Howell on Dribbble.